+++ js/main.js
@@ -1,5 +1,6 @@
 import {createOffers} from './create-offer.js';
 import {renderCard} from './render-card.js';
+import {validatePage} from './validation.js';
 
 const OFFER_AMOUNT = 10;
 
@@ -8,3 +9,4 @@ const map = document.querySelector('#map-canvas');
 const data = createOffers(OFFER_AMOUNT);
 const card = renderCard(data[0]);
 map.appendChild(card);
+validatePage();
